% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\docType{data}
\name{bc_albers}
\alias{bc_albers}
\title{BC Albers Equal Area Projection proj.4 string for British Columbia}
\format{An object of class \code{character} of length 1.}
\usage{
bc_albers
}
\description{
Shortcut that provides a proj.4 string
}
\keyword{datasets}
