% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\docType{data}
\name{etrs_laea}
\alias{etrs_laea}
\title{ETRS89 Lambert Azimuthal Equal Area Projection proj.4 string for Europe}
\format{An object of class \code{character} of length 1.}
\usage{
etrs_laea
}
\description{
Shortcut that provides a proj.4 string
}
\keyword{datasets}
