% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\docType{data}
\name{geo_wgs84}
\alias{geo_wgs84}
\title{Geographic, decimal degrees, wgs84 proj.4 string}
\format{An object of class \code{character} of length 1.}
\usage{
geo_wgs84
}
\description{
Shortcut that provides a proj.4 string
}
\keyword{datasets}
