% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\docType{data}
\name{usgs_albers}
\alias{usgs_albers}
\title{USGS Albers Equal Area Projection proj.4 string}
\format{An object of class \code{character} of length 1.}
\usage{
usgs_albers
}
\description{
Shortcut that provides a proj.4 string
}
\keyword{datasets}
