% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{utm_wgs84}
\alias{utm_wgs84}
\title{UTM zone projection with wgs84 proj.4 string}
\usage{
utm_wgs84(zone)
}
\arguments{
\item{zone}{chr string of UTM zone}
}
\description{
Shortcut that provides a proj.4 string
}
\examples{
utm_wgs84('10')
}
