% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\docType{data}
\name{web_mercator}
\alias{web_mercator}
\title{Web Mercator proj.4 string}
\format{An object of class \code{character} of length 1.}
\usage{
web_mercator
}
\description{
Shortcut that provides a proj.4 string
}
\keyword{datasets}
