% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{bc_albers}
\alias{bc_albers}
\alias{epsg_3005}
\title{BC Albers Equal Area Projection proj.4 string}
\usage{
bc_albers()

epsg_3005()
}
\description{
Function that provides a proj.4 string for the BC Albers projection (EPSG code 3005).
}
\details{
\href{http://epsg.io/3005}{BC Albers on epsg.io}
}
